+++ TP_[1]/src/calculos.h
@@ -2,7 +2,7 @@
  * calculos.h
  *
  *  Created on: 16 sep. 2022
- *      Author: braia
+ *      Author: braian
  */
 
 #ifndef CALCULOS_H_
@@ -10,6 +10,6 @@
 
 int calcularPromedio(int sumaCantidades, int divisor, float * promedio);
 int calcularPorcentaje(int cantidadTotal, int cantidadPromedio, float * porcentaje);
-int calcularInteres(int precio, int porcentajeInteres, float * aumento, float * precioConInteres);
+int calcularInteres(float precio, int porcentajeInteres, float * aumento, float * precioConInteres);
 
 #endif /* CALCULOS_H_ */
